+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - The `RagSystem.format_answer` now uses HTML `<br>` tags to separate newlines, to make
   it fully compatible to wrap in an HTML rendering context.
 
+### Fixed
+- Fixed a bug in `NumpyEmbeddingStore` when there were fewer than `num_docs` embeddings
+  in the store, causing an error when trying to retrieve embeddings.
+
 
 ## [v1.1.0] - 2024-08-13
 ### Added